Text Summarizer/requirements.txt
transformers
python-docx
fpdf
streamlit
pillow
requests
PyPDF2
beautifulsoup4
selenium
webdriver-manager